feat: 🎸 load 放在common 中， Component 删除掉load
