+++ packages/core/src/common/index.ts
@@ -5,3 +5,4 @@ export { default as zip } from './zip';
 export { default as unzip } from 'decompress';
 export { default as i18n } from './i18n';
 export { default as help } from './help';
+export { default as load } from './load';
