+++ packages/core/test/fixtures/load.ts
@@ -1,0 +1,3 @@
+import { load } from '../../src/common';
+
+load('fc@0.1.2', 'alibaba');
